Issue 203_2: MediaFileBase gone
