+++ app/src/main/resources/ApplicationResources.properties
@@ -715,10 +715,8 @@ maintenance.message.reset=Successfully reset the hit count of your Roller weblog
 # ----------------------------------------------------------- Media file management
 
 mediaFileImageChooser.subtitle=Choose an image
-mediaFileImageChooser.rootPageTip=Click to choose an image or click on a \
+mediaFileImageChooser.pageTip=Click to choose an image or click on a \
 directory to enter it.
-mediaFileImageChooser.dirPageTip=Click to choose an image or use the Switch to Folder \
-drop-down to move to another directory.
 
 
 # -- Media File Add
@@ -796,7 +794,6 @@ This page shows the media files that you have uploaded as well as files \
 copied in when you chose to customize a theme. You can use folders \
 to organize your files, and you do not have to worry about "breaking \
 URLs" when you move or rename files.
-mediaFileView.dirPageTip=This is a media file subfolder.
 
 mediaFileView.viewFolder=View Folder
